requirements.txt
streamlit==1.22.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.2.2
matplotlib==3.7.2
seaborn==0.12.2
rasa==3.5.14
rasa-sdk==3.5.1
plotly==5.15.0
altair==4.2.2
requests==2.31.0
joblib==1.3.2
flask==2.3.2
streamlit-chat==0.1.1
python-dotenv==0.21.0
sqlalchemy<2.0
streamlit-chat==0.1.1
